overwrote original tr-image with easy-tr
